refactor: Separate PRofileScreen into mutliple Composables corresponding to sections
