+++ app/src/main/java/com/example/keyraceapp/data/repositories/ScoreRepositoryImpl.kt
@@ -1,7 +1,6 @@
 package com.example.keyraceapp.data.repositories
 
 import com.example.keyraceapp.data.local.KeyRaceDatabase
-import com.example.keyraceapp.data.mappers.toArcadeScore
 import com.example.keyraceapp.data.mappers.toArcadeScoreEntity
 import com.example.keyraceapp.data.mappers.toTrainingScore
 import com.example.keyraceapp.data.mappers.toTrainingScoreEntity
@@ -41,15 +40,15 @@ class ScoreRepositoryImpl(
             try {
                 emit(Resource.Loading())
 
-                val result = arcadeScoreDao
+/*                val result = arcadeScoreDao
                     .getTopTenScores()
-                    .map { it.toArcadeScore()}
+                    .map { it.toArcadeScore()}*/
                 //MOCK DATA TO SEE THE UI
-/*                val result: List<Score.ArcadeScore> = listOf(
+                val result: List<Score.ArcadeScore> = listOf(
                     Score.ArcadeScore(wpm = 50f, acc = 100f, points = 300),
                     Score.ArcadeScore(wpm = 20f, acc = 100f, points = 160),
                     Score.ArcadeScore(wpm = 10f, acc = 98f, points = 160),
-                )*/
+                )
